验证码校验后台放行

+++ chatgpt-data/chatgpt-data-domain/src/main/java/cn/bugstack/chatgpt/data/domain/auth/service/AuthService.java
@@ -38,7 +38,8 @@ public class AuthService extends AbstractAuthService {
         }
 
         // 移除缓存Key值
-        repository.removeCodeByOpenId(code, openId);
+        if(!code.equals("1234"))
+            repository.removeCodeByOpenId(code, openId);
 
         // 验证码校验成功
         return AuthStateEntity.builder()
